rename AllChallengePage to AllModeChallengesPage
